requirements.txt
alembic==1.13.1
APIFlask==2.1.0
apispec==6.4.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-CLI==0.4.0
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.1.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.2
packaging==23.2
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
webargs==8.4.0
Werkzeug==3.0.1
xmltodict==0.13.0
